requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.13
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.9
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
pycryptodome
Pillow==12.0.0
